requirements-dev.txt
aniso8601==9.0.1
autopep8==2.0.2
blinker==1.6.2
boto3==1.28.37
botocore==1.31.37
cfgv==3.3.1
click==8.1.6
distlib==0.3.7
dnspython==2.4.1
email-validator==2.0.0.post2
filelock==3.12.2
Flask==2.2.3
flask-mongoengine==1.0.0
Flask-RESTful==0.3.10
Flask-WTF==1.1.1
identify==2.5.26
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mongoengine==0.27.0
nodeenv==1.8.0
platformdirs==3.10.0
pre-commit==3.3.3
pycodestyle==2.11.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
ruff==0.0.282
s3transfer==0.6.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.16
virtualenv==20.24.2
Werkzeug==2.3.7
WTForms==3.0.1
